Merge master to 4-factor branch

+++ src/basis/basis_bernstein.c
@@ -15,9 +15,34 @@ interp_result_t bernstein_basis_create(basis_set_t **out, const basis_spec_t spe
         return INTERP_ERROR_FAILED_ALLOCATION;
 
     bernstein_interpolation_value_derivative_matrix(rule->n_nodes, integration_rule_nodes_const(rule), spec.order,
-                                                    this->_data, this->_data + (spec.order + 1) * rule->n_nodes);
+                                                    this->_data,
+                                                    this->_data + (size_t)(spec.order + 1) * rule->n_nodes);
+
     this->integration_spec = rule->spec;
     this->spec = spec;
     *out = this;
+
+    // printf("Computed Bernstein basis of order %u on rule with %u nodes:\n", spec.order, rule->n_nodes);
+    // for (unsigned i_basis = 0; i_basis < spec.order + 1; ++i_basis)
+    // {
+    //     printf("\tBasis %u:", i_basis);
+    //     const double *basis = basis_set_basis_values(this, i_basis);
+    //     for (unsigned ipt = 0; ipt < rule->n_nodes; ++ipt)
+    //         printf(" %5.2g", basis[ipt]);
+    //     printf("\n");
+    // }
+    // for (unsigned i_basis = 0; i_basis < spec.order + 1; ++i_basis)
+    // {
+    //     printf("\tDerivative %u:", i_basis);
+    //     const double *derivative = basis_set_basis_derivatives(this, i_basis);
+    //     for (unsigned ipt = 0; ipt < rule->n_nodes; ++ipt)
+    //         printf(" %5.2g", derivative[ipt]);
+    //     printf("\n");
+    // }
+    // printf("\t Raw data:");
+    // for (unsigned ipt = 0; ipt < rule->n_nodes * 2 * (spec.order + 1); ++ipt)
+    //     printf(" %5.2g", this->_data[ipt]);
+    // printf("\n");
+
     return INTERP_SUCCESS;
 }
